LayoutTests/imported/w3c/web-platform-tests/trusted-types/require-trusted-types-for-TypeError-belongs-to-the-global-object-realm-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"


PASS Setting innerHTML on a node inserted by the parser.
FAIL Setting innerHTML on a node adopted from a subframe. assert_throws_js: function "_ => divAdoptedFromIframe.innerHTML = 'unsafe'" threw object "TypeError: This assignment requires a TrustedHTML" ("TypeError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
